fix(purge): change required_excess parameter type from int to double
The required_excess parameter accepts floating-point values (e.g. 0.02, 0.09, 0.0)
in all parameter files and in Python scripts, but was declared as `int` in
filter/purge.cpp. Passing a non-integer value (or even the default 0.0) caused:

  parameter_error: cannot cast parameter required_excess to type i:
    incorrect double-to-integer conversion in parameter

Fix by changing the parameter type from `int` to `double` to match the
actual semantics — required_excess represents a fractional excess ratio,
consistently used as a double throughout the codebase.

+++ filter/purge.cpp
@@ -129,7 +129,7 @@ struct purge_process : purge_output_specification {
         "-1">
         nsteps;
     parameter_with_default<
-        int, "required_excess",
+        double, "required_excess",
         R"(% of excess required at the end of the 1st singleton removal step)",
         CADO_STRINGIZE(DEFAULT_PURGE_REQUIRED_EXCESS)>
         required_excess;
